requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
psycopg2-binary>=2.9.1
pydantic>=1.8.2
pydantic-settings>=2.0.0
python-dotenv>=0.19.0
email-validator>=1.1.3
python-multipart>=0.0.5
aiofiles>=0.7.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=3.2.0
pillow>=8.3.2
reportlab>=3.6.1
beautifulsoup4>=4.9.3
requests>=2.26.0
dnspython>=2.1.0
pytz>=2021.1
